@@ -1,6 +1,6 @@
 module github.com/jfcg/sorty/v2
 
-go 1.17
+go 1.18
 
 require (
 	github.com/jfcg/opt v0.3.1
